*move MergeThreeConvolutions, MergeTwoConvolutions, MergeParallelConvolutions, MergeConcatedParallelConvolutions, MergeParallelDepthwiseConvolutions and MergeParallelScaleAndDepthwiseConvolution to separate source files.
