+++ prj/vs2022/CvtCore.vcxproj
@@ -27,7 +27,7 @@
     <ClCompile Include="..\..\src\Cvt\Deoptimizer\Deoptimizer.cpp" />
     <ClCompile Include="..\..\src\Cvt\InferenceEngine\InferenceEngine.cpp" />
     <ClCompile Include="..\..\src\Cvt\Optimizer\MergeBiasAndScale.cpp" />
-    <ClCompile Include="..\..\src\Cvt\Optimizer\MergeConvolution.cpp" />
+    <ClCompile Include="..\..\src\Cvt\Optimizer\MergeConcatedParallelConvolutions.cpp" />
     <ClCompile Include="..\..\src\Cvt\Optimizer\MergeConvolutionAndPower.cpp" />
     <ClCompile Include="..\..\src\Cvt\Optimizer\MergeConvolutionAndScale.cpp" />
     <ClCompile Include="..\..\src\Cvt\Optimizer\MergeConvolutionOrOtherAndActivation.cpp" />
@@ -38,6 +38,9 @@
     <ClCompile Include="..\..\src\Cvt\Optimizer\MergeInnerProductAndScale.cpp" />
     <ClCompile Include="..\..\src\Cvt\Optimizer\MergeMish.cpp" />
     <ClCompile Include="..\..\src\Cvt\Optimizer\MergeNormalize.cpp" />
+    <ClCompile Include="..\..\src\Cvt\Optimizer\MergeParallelConvolutions.cpp" />
+    <ClCompile Include="..\..\src\Cvt\Optimizer\MergeParallelDepthwiseConvolutions.cpp" />
+    <ClCompile Include="..\..\src\Cvt\Optimizer\MergeParallelScaleAndDepthwiseConvolution.cpp" />
     <ClCompile Include="..\..\src\Cvt\Optimizer\MergePermute.cpp" />
     <ClCompile Include="..\..\src\Cvt\Optimizer\MergePooling.cpp" />
     <ClCompile Include="..\..\src\Cvt\Optimizer\MergePowerAndScaleAndPower.cpp" />
@@ -51,7 +54,9 @@
     <ClCompile Include="..\..\src\Cvt\Optimizer\MergeSpaceToDepth.cpp" />
     <ClCompile Include="..\..\src\Cvt\Optimizer\MergeSqueezeExcitation.cpp" />
     <ClCompile Include="..\..\src\Cvt\Optimizer\MergeSwish.cpp" />
+    <ClCompile Include="..\..\src\Cvt\Optimizer\MergeThreeConvolutions.cpp" />
     <ClCompile Include="..\..\src\Cvt\Optimizer\MergeTiledScale2D.cpp" />
+    <ClCompile Include="..\..\src\Cvt\Optimizer\MergeTwoConvolutions.cpp" />
     <ClCompile Include="..\..\src\Cvt\Optimizer\MergeUnpack.cpp" />
     <ClCompile Include="..\..\src\Cvt\Optimizer\MergeYoloV7.cpp" />
     <ClCompile Include="..\..\src\Cvt\Optimizer\Optimizer.cpp" />
